+++ gcc/rust/backend/rust-compile-base.h
@@ -18,6 +18,7 @@
 #define RUST_COMPILE_BASE
 
 #include "rust-compile-context.h"
+#include "rust-compile-type.h"
 #include "rust-hir-visitor.h"
 #include "rust-hir-full.h"
